+++ docs/CHANGELOG.md
@@ -1,9 +1,21 @@
+* HttpService file upload fixed
+* Added option to set on/off values for switch
+* Added ability to indicate errors in Language Input
+* Added language validation rule
+
+<a name="0.0.2-beta.3"></a>
+## 0.0.2-beta.3 (2016-08-30)
+
+* Updated validation error indicator
+* Fixed datagrid layout issue
+* Updated stylings for button group and switch
+
+
 <a name="0.0.1-beta.2"></a>
 ## 0.0.1-beta.2 (2016-08-29)
 
 * Added support to use em/rem values for Dialog width/height and Data-Column widths
 * Updated validation-renderer
-
 
 
 <a name="0.0.1-beta.1"></a>